Improve dependency injection to move it out of the class definitions
