+++ src/application/service/CreateNewMagicLink.ts
@@ -1,9 +1,7 @@
 import { MagicLink } from '../../domain/model/magiclink/MagicLink';
 import { v4 } from 'uuid';
 import { Email } from '../../domain/model/user/Email';
-import magicLinkRepository, {
-  MagicLinkRepository,
-} from '../../domain/model/magiclink/MagicLinkRepository';
+import { MagicLinkRepository } from '../../domain/model/magiclink/MagicLinkRepository';
 
 export class CreateNewMagicLink {
   constructor(private magicLinkRepository: MagicLinkRepository) {}
@@ -15,5 +13,3 @@ export class CreateNewMagicLink {
     return this.magicLinkRepository.save(magicLink);
   }
 }
-
-export const createMagicLink = new CreateNewMagicLink(magicLinkRepository);
