--- src/domain/model/magiclink/__mocks__/MagicLinkRepository.ts
@@ -1,4 +1,0 @@
-export default {
-  save: jest.fn(),
-  findByCode: jest.fn(),
-};
