+++ src/infrastructure/persistence/index.ts
@@ -1,0 +1,4 @@
+import database from '../../database';
+import { PsqlMagicLinkRepository } from './PsqlMagicLinkRepository';
+
+export const magicLinkRepository = new PsqlMagicLinkRepository(database);
